+++ CPP06/ex00/ScalarConverterIdentityUtils.cpp
@@ -1,8 +1,8 @@
 #include "ScalarConverter.hpp"
 
-unsigned int ScalarConverter::len_skipping_zero(
+unsigned long ScalarConverter::len_skipping_zero(
     const std::string literal) {
-  unsigned int i = 0;
+  unsigned long i = 0;
   if (literal.at(i) == '+' || literal.at(i) == '-') i++;
   while (i < literal.length() && literal.at(i) == '0') i++;
   return (literal.length() - i);
